+++ CHANGELOG.md
@@ -5,6 +5,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 
+### Changed
+- Fail cleanly if exception occurs while accessing subscription event source
+
 ## [2.4.8] - 2020-01-29
 ### Changed
 - Inserts into the event-buffer no longer fails if there is a conflict; it just logs a warning